Move shm functions. Fix mmap() length bug.
Move shm functions to their own file, and provide interface in shared
header. This way server and clients can use the same code. Fix a bug in
call to mmap() where `length` was being passed as 0. It should be the
length of the mapping you want. Also changed the nature of
`daemonize()` to allow the server to print to stdout, for debugging.