Update client/server sync to use mutexes/condvars.
Update the registrar() and supporting functions to use condvars and
mutexes to synchronize work between the client and server. Add
supporting functions to initialize the ringbuffer (using
PTHREAD_PROCESS_SHARED for the mutexes/condvars) and to handle
individual client requests. Currently all functions are specific to
registration ring buffers, but they should be extendable to file-service
ring buffers.