Push all request-handling functionality into macros.
With this commit, all request handling is taken care of through macros
and a single function pointer: the function must take in a pointer to
a sring_entry (containing the request in entry->req and space for the
response), process the request, and push the reponse back to entry->rsp.
This is the only buffer-type-specific function on the server side. Also
it takes in an expression "done" which breaks the infinite loop when it
evaluates to True.

So now the whole process goes, for the server:
    struct <tag>_sring *r = shm_create(fname, sizeof(*r));
    RB_INIT(<tag>, r, SLOT_COUT);
    RB_SERVE(<tag>, r, done, &handler_func);
    shm_destroy(fname, r, size);

And for the client it is even simpler:
    struct <tag>_sring *r = shm_map(fname, size);
    [[ setup request and response structs ]]
    RB_MAKE_REQUEST(<tag>, r, &request, &response);
    shm_unmap(fname, sizeof(*r));